[Feat] 책임 분리에 따른 파일 구조 수정 및 의존성 추가
