+++ mokakbob-core/build.gradle
@@ -1,11 +1,3 @@
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-
-    // smtp
-    implementation 'org.springframework.boot:spring-boot-starter-mail'
-
-    // jwt
-    implementation("io.jsonwebtoken:jjwt-api:0.11.5")
-    runtimeOnly("io.jsonwebtoken:jjwt-impl:0.11.5")
-    runtimeOnly("io.jsonwebtoken:jjwt-jackson:0.11.5")
 }
